forum topic decomposition;import translation bug;
